KS-2-1 third refactoring

+++ src/main/kotlin/lesson_2/KS-2-1.kt
@@ -2,13 +2,13 @@ package org.example.lesson_2
 
 fun main() {
 
-    val students = 4.0
+    val students = 4
     val firstStudent = 3
     val secondStudent = 4
     val thirdStudent = 3
     val fourthStudent = 5
 
-    val average = (firstStudent + secondStudent + thirdStudent + fourthStudent) / students
+    val average = (firstStudent + secondStudent + thirdStudent + fourthStudent) / students.toFloat()
 
     println(average)
